clerk for auth and tiptap as the new editor.
