+++ package.json
@@ -12,52 +12,57 @@
 		"lint": "prettier --check ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^5.0.3",
-		"autoprefixer": "^10.4.20",
-		"daisyui": "^5.0.35",
-		"globals": "^15.0.0",
-		"prettier": "^3.3.2",
-		"prettier-plugin-svelte": "^3.2.6",
-		"prettier-plugin-tailwindcss": "^0.6.5",
-		"svelte": "^5.16.0",
-		"svelte-check": "^4.0.0",
-		"tailwindcss": "^3.4.9",
-		"typescript": "^5.0.0",
-		"vite": "^6.0.6",
-		"vite-plugin-static-copy": "^2.2.0"
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.26.1",
+		"@sveltejs/vite-plugin-svelte": "^5.1.1",
+		"autoprefixer": "^10.4.21",
+		"daisyui": "^5.0.47",
+		"globals": "^15.15.0",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"prettier-plugin-tailwindcss": "^0.6.14",
+		"svelte": "^5.36.16",
+		"svelte-check": "^4.3.0",
+		"tailwindcss": "^3.4.17",
+		"typescript": "^5.8.3",
+		"vite": "^6.3.5",
+		"vite-plugin-static-copy": "^2.3.1"
 	},
 	"dependencies": {
-		"@auth0/auth0-spa-js": "^2.1.3",
+		"@auth0/auth0-spa-js": "^2.3.0",
 		"@excalidraw/excalidraw": "^0.17.6",
 		"@friendofsvelte/tipex": "^0.0.8",
-		"@iconify/svelte": "5",
+		"@iconify/svelte": "^5.0.1",
 		"@neondatabase/serverless": "^0.10.4",
 		"@tailwindcss/container-queries": "^0.1.1",
 		"@tailwindcss/forms": "^0.5.10",
 		"@tailwindcss/typography": "^0.5.16",
-		"@tinymce/tinymce-svelte": "^3.0.0",
-		"@tiptap/extension-heading": "^2.22.2",
-		"@tiptap/extension-highlight": "^2.22.3",
-		"@tiptap/extension-subscript": "^2.22.3",
-		"@tiptap/extension-table": "^2.22.3",
-		"@tiptap/extension-table-cell": "^2.22.3",
-		"@tiptap/extension-table-header": "^2.22.3",
-		"@tiptap/extension-table-row": "^2.22.3",
-		"@tiptap/extension-text-align": "^2.22.3",
-		"@tiptap/extension-typography": "^2.22.3",
-		"@tiptap/extension-underline": "^2.22.2",
-		"@tiptap/extension-youtube": "^2.22.3",
+		"@tinymce/tinymce-svelte": "^3.1.0",
+		"@tiptap/core": "^3.0.7",
+		"@tiptap/extension-heading": "^3.0.7",
+		"@tiptap/extension-highlight": "^2.26.1",
+		"@tiptap/extension-mathematics": "^3.0.7",
+		"@tiptap/extension-subscript": "^2.26.1",
+		"@tiptap/extension-table": "^2.26.1",
+		"@tiptap/extension-table-cell": "^2.26.1",
+		"@tiptap/extension-table-header": "^2.26.1",
+		"@tiptap/extension-table-row": "^2.26.1",
+		"@tiptap/extension-text-align": "^2.26.1",
+		"@tiptap/extension-typography": "^2.26.1",
+		"@tiptap/extension-underline": "^2.26.1",
+		"@tiptap/extension-youtube": "^2.26.1",
+		"@tiptap/pm": "^3.0.7",
+		"@tiptap/starter-kit": "^3.0.7",
 		"cookie": "^1.0.2",
 		"dompurify": "^3.2.6",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
+		"katex": "^0.16.22",
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
 		"string-hash": "^1.1.3",
 		"svelte-clerk": "^0.13.4",
 		"svelte-toasts": "^1.1.2",
-		"tinymce": "^7",
-		"uuid": "^11.0.3"
+		"tinymce": "^7.9.1",
+		"uuid": "^11.1.0"
 	},
 	"packageManager": "pnpm@10.13.1"
 }
