+++ src/routes/components/tiptap-editor.css
@@ -80,26 +80,6 @@ table th {
     border-radius: 4px;
 }
 
-/* Add tooltips for the buttons */
-.tipex-edit-button {
-    position: relative;
-}
-
-.tipex-edit-button:hover::after {
-    content: attr(title);
-    position: absolute;
-    bottom: 100%;
-    left: 50%;
-    transform: translateX(-50%);
-    padding: 4px 8px;
-    background-color: #374151;
-    color: white;
-    border-radius: 4px;
-    font-size: 12px;
-    white-space: nowrap;
-    z-index: 10;
-}
-
 /* Table styles */
 table {
     border-collapse: collapse;
